Extract PSQLRow (#177)
### Motivation

In #135, I did some naming things just wrong. `PSQLRows` is a stupid name. We should name it `PSQLRowStream`. `PSQLRows.Row` is a stupid name for a single table row. We should name it `PSQLRow`. Transforming an incoming data row packet to a `[PSQLData]` array early is expensive and stupid. Let's not do this anymore.

### Changes

- Extract `PSQLRows.Row` to `PSQLRows` (Got its own file). Stop the early `[PSQLData]` madness.
- Rename `PSQLRows` to `PSQLRowStream`
- Fix naming in integration tests to match `PSQLRowStream`