`close()` closes immediately; Add new `closeGracefully()` (#383)
Fixes #370.

This patch changes the behavior of `PostgresConnection.close()`. Currently `close()` terminates the connection only after all queued queries have been successfully processed by the server. This however leads to an unwanted dependency on the Postgres server to close a connection. If a server stops responding, the client is currently unable to close its connection. Because of this, this patch changes the behavior of `close()`. `close()` now terminates a connection immediately and fails all running or queued queries. To allow users to continue to use the existing behavior we introduce a `closeGracefully()` that now has the same behavior as close had previously. 

Since we never documented the old close behavior and we consider it dangerous in certain situations we are fine with changing the behavior without tagging a major version.